refactor(programa-formacion): rediseñando y aplicando livewire a las vistas

+++ database/migrations/batch_16_inventario/2025_08_12_184612_create_productos_table.php
@@ -11,28 +11,30 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('productos', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('productos')) {
+            Schema::create('productos', function (Blueprint $table) {
+                $table->id();
 
-            $table->string('producto');
-            $table->unsignedBigInteger('tipo_producto_id');
-            $table->text('descripcion');
-            $table->unsignedBigInteger('unidad_medida_id');
-            $table->unsignedBigInteger('estado_id');
+                $table->string('producto');
+                $table->unsignedBigInteger('tipo_producto_id');
+                $table->text('descripcion');
+                $table->unsignedBigInteger('unidad_medida_id');
+                $table->unsignedBigInteger('estado_id');
 
-            $table->unsignedBigInteger('user_create_id');
-            $table->unsignedBigInteger('user_update_id');
+                $table->unsignedBigInteger('user_create_id');
+                $table->unsignedBigInteger('user_update_id');
 
-            $table->timestamps();
+                $table->timestamps();
 
-            //Se relacionan las llaves foráneas
-            $table->foreign('tipo_producto_id')->references('id')->on('parametros_temas')->onDelete('restrict');
-            $table->foreign('unidad_medida_id')->references('id')->on('parametros_temas')->onDelete('restrict');
-            $table->foreign('estado_id')->references('id')->on('parametros_temas')->onDelete('restrict');
-            
-            $table->foreign('user_create_id')->references('id')->on('users')->onDelete('restrict');
-            $table->foreign('user_update_id')->references('id')->on('users')->onDelete('restrict');
-        });
+                //Se relacionan las llaves foráneas
+                $table->foreign('tipo_producto_id')->references('id')->on('parametros_temas')->onDelete('restrict');
+                $table->foreign('unidad_medida_id')->references('id')->on('parametros_temas')->onDelete('restrict');
+                $table->foreign('estado_id')->references('id')->on('parametros_temas')->onDelete('restrict');
+                
+                $table->foreign('user_create_id')->references('id')->on('users')->onDelete('restrict');
+                $table->foreign('user_update_id')->references('id')->on('users')->onDelete('restrict');
+            });
+        }
     }
     /**
      * Reverse the migrations.
